+++ testsuite/tests/concurrent/should_run/T3279.hs
@@ -7,8 +7,15 @@ import GHC.IO (unsafeUnmask)
 
 f :: Int
 f = (1 +) . unsafePerformIO $ do
-        throwIO (ErrorCall "foo") `catch` \(SomeException e) -> do
-            myThreadId >>= flip throwTo e
+        -- NB: We have to use `catchNoPropagate` rather than `catch`.
+        -- `catch` starting from !13302 turns all asynchronous exceptions into synchronous ones
+        -- because it has to catch and rethrow the exception to annotate it
+        -- with the `WhileHandling` backtrace. Using `catchNoPropagate`
+        -- preserves the asynchronous exception stack unwinding behaviour which
+        -- allows this test to work.
+        -- See #25300 and !13302 for details.
+        throwIO (ErrorCall "foo") `catchNoPropagate` \(ExceptionWithContext c (SomeException e)) -> do
+            myThreadId >>= flip throwTo (ExceptionWithContext c (SomeException e))
             -- point X
             unsafeUnmask $ return 1